net: mqtt_sn: widen gateway duration to avoid sentinel collision
mqtt_sn_gw_create() took the advertised duration as a short, but ADVERTISE
carries a uint16_t duration. A gateway advertising duration 0xFFFF
narrowed to (short)-1, which the function treats as the 'never expires'
sentinel, so that gateway would never be timed out; durations >= 0x8000
became negative and yielded an already-expired adv_timer. The update path
uses p->duration (uint16_t) directly, so the two paths disagreed.

Widen the parameter to int so the -1 sentinel (passed by the SEARCHGW /
GWINFO callers) and the full 0..65535 wire range are unambiguous.

+++ subsys/net/lib/mqtt_sn/mqtt_sn.c
@@ -359,7 +359,7 @@ static void mqtt_sn_gw_destroy_all(struct mqtt_sn_client *client)
 	}
 }
 
-static struct mqtt_sn_gateway *mqtt_sn_gw_create(uint8_t gw_id, short duration,
+static struct mqtt_sn_gateway *mqtt_sn_gw_create(uint8_t gw_id, int duration,
 						 struct mqtt_sn_data gw_addr)
 {
 	struct mqtt_sn_gateway *gw;
